+++ src/main/java/core/jdbc/PreparedStatementSetter.java
@@ -1,0 +1,8 @@
+package core.jdbc;
+
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+
+public interface PreparedStatementSetter {
+    void setValues(PreparedStatement pstmt) throws SQLException;
+}
