+++ src/main/java/core/jdbc/RowMapper.java
@@ -1,0 +1,8 @@
+package core.jdbc;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+public interface RowMapper {
+    Object mapRow(ResultSet rs) throws SQLException;
+}
