+++ src/main/java/next/dao/UserDao.java
@@ -1,80 +1,48 @@
 package next.dao;
 
+import core.jdbc.JdbcTemplate;
 import next.model.User;
 
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
 import java.util.stream.Collectors;
 
 public class UserDao {
-
     public void insert(User user) throws SQLException {
-        final JdbcTemplate jdbcTemplate = new JdbcTemplate() {
-            @Override
-            protected void setValues(PreparedStatement pstmt) throws SQLException {
-                pstmt.setString(1, user.getUserId());
-                pstmt.setString(2, user.getPassword());
-                pstmt.setString(3, user.getName());
-                pstmt.setString(4, user.getEmail());
-            }
-
-            @Override
-            protected Object mapRow(ResultSet rs) {
-                return null;
-            }
-        };
-        jdbcTemplate.update("INSERT INTO USERS VALUES (?, ?, ?, ?)");
+        final JdbcTemplate jdbcTemplate = new JdbcTemplate();
+        jdbcTemplate.update("INSERT INTO USERS VALUES (?, ?, ?, ?)", pstmt -> {
+            pstmt.setString(1, user.getUserId());
+            pstmt.setString(2, user.getPassword());
+            pstmt.setString(3, user.getName());
+            pstmt.setString(4, user.getEmail());
+        });
     }
 
     public void update(User user) throws SQLException {
-        final JdbcTemplate jdbcTemplate = new JdbcTemplate() {
-            @Override
-            protected void setValues(PreparedStatement pstmt) throws SQLException {
-                pstmt.setString(1, user.getName());
-                pstmt.setString(2, user.getEmail());
-                pstmt.setString(3, user.getUserId());
-            }
-
-            @Override
-            protected Object mapRow(ResultSet rs) {
-                return null;
-            }
-        };
-        jdbcTemplate.update("UPDATE USERS SET name = ?, email = ? WHERE userId = ?");
+        final JdbcTemplate jdbcTemplate = new JdbcTemplate();
+        jdbcTemplate.update("UPDATE USERS SET name = ?, email = ? WHERE userId = ?", pstmt -> {
+            pstmt.setString(1, user.getName());
+            pstmt.setString(2, user.getEmail());
+            pstmt.setString(3, user.getUserId());
+        });
     }
 
     public List<User> findAll() throws SQLException {
-        final JdbcTemplate jdbcTemplate = new JdbcTemplate() {
-            @Override
-            protected void setValues(PreparedStatement pstmt) {
-            }
-
-            @Override
-            protected User mapRow(ResultSet rs) throws SQLException {
-                return new User(rs.getString("userId"), rs.getString("password"), rs.getString("name"),
-                        rs.getString("email"));
-            }
-        };
-        return jdbcTemplate.query("SELECT userId, password, name, email FROM USERS").stream()
+        final JdbcTemplate jdbcTemplate = new JdbcTemplate();
+        final List<Object> list = jdbcTemplate.query("SELECT userId, password, name, email FROM USERS", pstmt -> {
+                },
+                rs -> new User(rs.getString("userId"), rs.getString("password"), rs.getString("name"),
+                        rs.getString("email")));
+        return list.stream()
                 .map(o -> (User) o)
                 .collect(Collectors.toList());
     }
 
     public User findByUserId(String userId) throws SQLException {
-        final JdbcTemplate jdbcTemplate = new JdbcTemplate() {
-            @Override
-            protected void setValues(PreparedStatement pstmt) throws SQLException {
-                pstmt.setString(1, userId);
-            }
-
-            @Override
-            protected User mapRow(ResultSet rs) throws SQLException {
-                return new User(rs.getString("userId"), rs.getString("password"), rs.getString("name"),
-                        rs.getString("email"));
-            }
-        };
-        return (User) jdbcTemplate.queryForObject("SELECT userId, password, name, email FROM USERS WHERE userid = ?");
+        final JdbcTemplate jdbcTemplate = new JdbcTemplate();
+        return (User) jdbcTemplate.queryForObject("SELECT userId, password, name, email FROM USERS WHERE userid = ?",
+                pstmt -> pstmt.setString(1, userId),
+                rs -> new User(rs.getString("userId"), rs.getString("password"), rs.getString("name"),
+                        rs.getString("email")));
     }
 }
